refactor(settings/recharge): 提取 PointsDashboard 为独立组件
