+++ src/renderer/pages/settings/recharge/index.tsx
@@ -6,6 +6,7 @@ import { useAuth } from '@/renderer/context/AuthContext';
 import { ipcBridge } from '@/common';
 import PageWrapper from '@renderer/components/base/PageWrapper';
 import OrderList from './components/OrderList';
+import PointsDashboard from './components/PointsDashboard';
 import { OrderStatusEnum } from './types';
 import type { CreateOrderResponse, OrderStatus, PaymentMethod, PayOrderResponse, RechargePackage, RechargeStep } from './types';
 import { formatCurrency } from './utils';
@@ -16,37 +17,7 @@ const QRCodeSVGLazy = React.lazy(async () => {
   return { default: mod.QRCodeSVG };
 });
 
-// ==================== Component ====================
-
-// 充值页通用「面板」容器样式 / Shared panel container style
-const PANEL_CLASS = 'p-6 bg-muted rd-16px border border-solid border-[var(--border-light)]';
-
-const PointsDashboard: React.FC<{ remainingPoints: number; usedPoints: number; bonusPoints: number }> = ({ remainingPoints, usedPoints, bonusPoints }) => {
-  const { t } = useTranslation();
-
-  const statItems = [
-    { label: t('settings.recharge.remainingPoints') || '剩余积分', value: remainingPoints, valueClass: 'italic text-[var(--ui-accent-orange)]' },
-    { label: t('settings.recharge.usedPoints') || '累计已用', value: usedPoints, valueClass: 'text-foreground' },
-    { label: t('settings.recharge.bonusPoints') || '赠送积分', value: bonusPoints, valueClass: 'text-foreground' },
-  ];
-
-  return (
-    <div className={PANEL_CLASS}>
-      <div className='text-14px font-600 text-foreground mb-6'>{t('settings.recharge.pointsInfo') || '积分信息'}</div>
-      <div className='grid grid-cols-3'>
-        {statItems.map(({ label, value, valueClass }) => (
-          <div key={label} className={`flex flex-col gap-2`}>
-            <div className='text-13px font-500 text-secondary'>{label}</div>
-            <div className='flex items-baseline gap-2'>
-              <span className={`text-32px font-700 leading-none ${valueClass}`}>{value.toLocaleString()}</span>
-              <span className='text-12px font-600 text-secondary'>PTS</span>
-            </div>
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-};
+const PANEL_CLASS = 'p-6 bg-muted rd-16px border border-light';
 
 const RechargeCenter: React.FC = () => {
   const { t } = useTranslation();
